Rename common deployment test classes -> mocks
