+++ python/ray/llm/tests/serve/cpu/integration/deployments/llm/multiplex/test_lora_deployment_base_client.py
@@ -13,8 +13,8 @@ from ray.llm._internal.serve.deployments.routers.router import (
     LLMRouter,
 )
 from ray.llm._internal.serve.configs.server_models import ModelData
-from ray.llm.tests.serve.deployments.fake_image_retriever import FakeImageRetriever
-from ray.llm.tests.serve.deployments.mock_vllm_engine import MockEchoVLLMEngine
+from ray.llm.tests.serve.mocks.fake_image_retriever import FakeImageRetriever
+from ray.llm.tests.serve.mocks.mock_vllm_engine import MockEchoVLLMEngine
 
 VLLM_APP_DEF = """
 model_loading_config:
