+++ python/ray/llm/tests/serve/mocks/fake_image_retriever.py
